+++ CHANGELOG.md
@@ -24,6 +24,10 @@ Main (unreleased)
 
 - Added `send_traceparent` option for `tracing` config to enable traceparent header propagation. (@MyDigitalLife)
 
+- Add `delay` option to `prometheus.exporter.cloudwatch` component to delay scraping of metrics to account for CloudWatch ingestion latency. (@tmeijn)
+
+- Export `yace_.*` metrics from the underlying YACE Exporter to `prometheus.exporter.cloudwatch`. (@tmeijn)
+
 - (_Public Preview_) Additions to `database_observability.mysql` and `database_observability.postgres` components:
   - `explain_plans`
     - always send an explain plan log message for each query, even skipped or errored queries. (@rgeyer)